refactor: restructure svc folder structure
